requirement.txt
asgiref==3.8.1
Django==5.0.7
django-mdeditor==0.1.20
Markdown==3.6
Pygments==2.18.0
python-markdown-math==0.8
sqlparse==0.5.1
tzdata==2024.1
